requirement.txt
fastapi
uvicorn
langchain
langchain-core
langchain-chroma
langchain-huggingface
langchain-deepseek
chromadb
sentence-transformers
python-dotenv
